Upload table generator
With this model the nominal current of a certain temperature limit can be calculated for multiple dimensions of the same conductor at once. These dimensions must be listed in excel and imported with the Import_from_excel function. The conductor can be selected in Properties_fun and Fun_total outputs the result.